v0.39beta
*Reverted to Azilroka's skin file after his latest update in AddOnSkins
(v2.15). A bit more fine tuning may needed.
*Removed the font.lua file. No need to force users to use BenikUI. On
it's final release, BenikUI will have a seperate addon to override ElvUI
name and damage font, to fit
with BenikUI. What was added in 0.37b wasn't working as intended.
*Added option for the actionbars switch buttons. The user can now choose
on which bar (1 or 2) the Switch buttons will show.
*Added style in WorldMap (needs testing for taints). The style applies
only on small and quest map not on the big map.
*Fixed a visibility bug when the game menu and the bags were opened.
*Fine tune the install profile.